Conflicts:
	uasdm-server/src/main/java-gen/stub/gov/geoplatform/uasdm/ImageryProcessingJob.java

+++ uasdm-server/src/main/java/gov/geoplatform/uasdm/geoserver/ImageMosaicService.java
@@ -1,17 +1,17 @@
 /**
  * Copyright 2020 The Department of Interior
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
  */
 package gov.geoplatform.uasdm.geoserver;
 
@@ -93,11 +93,14 @@ public class ImageMosaicService
       // Delete the existing store files
       final File[] files = baseDir.listFiles();
 
-      for (File file : files)
+      if (files != null)
       {
-        if (!file.isDirectory() && ( file.getName().startsWith(STORE_NAME) || file.getName().equals("sample_image.dat") ))
+        for (File file : files)
         {
-          FileUtils.deleteQuietly(file);
+          if (!file.isDirectory() && ( file.getName().startsWith(STORE_NAME) || file.getName().equals("sample_image.dat") ))
+          {
+            FileUtils.deleteQuietly(file);
+          }
         }
       }
     }
@@ -185,11 +188,14 @@ public class ImageMosaicService
   {
     final File[] files = baseDir.listFiles();
 
-    for (File file : files)
+    if (files != null)
     {
-      if (file.isDirectory())
+      for (File file : files)
       {
-        return true;
+        if (file.isDirectory())
+        {
+          return true;
+        }
       }
     }
